Create a Polish localisation
I used longer names for sand and glass to expand their purpose. However, glass is not made of ordinary sand but special for the production of a given type of glass.

Glass is also different. Here we have laboratory glass for chemical flasks.

And finally:
`Science` - I used a Polish synonym for science: knowledge. Knowledge creates Science.